move BookKeeperBuilder to impl package
