+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/client/LedgerCreateOp.java
@@ -105,7 +105,6 @@ class LedgerCreateOp implements GenericCallback<Void> {
      * Initiates the operation
      */
     public void initiate() {
-
         // allocate ensemble first
 
         /*
