+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/client/api/OpenBuilder.java
@@ -42,7 +42,7 @@ public interface OpenBuilder extends OpBuilder<ReadHandle> {
 
     /**
      * Define to open the ledger in recovery mode or in readonly mode. In recovery mode the ledger will be fenced and
-     * the writer of the ledger will be prevented from issuing other writes to the ledger. It defaults to 'true'
+     * the writer of the ledger will be prevented from issuing other writes to the ledger. It defaults to 'false'
      *
      * @param recovery recovery mode
      *
